Refactored the convert function
Moving parts of the code into a more iterative function that looks at
powers of 10.

We now have covered the "multiples of powers of 10" and
the "power of 10 minus previous power of 10" case.

Next up: "half of the power of 10" and "half of the power of 10 minus
the previous power of 10".